split snapshotting tests
